Unify algebraic tests for Linear, Polynomial, and Quadratic
