+++ rust/ommx/src/convert.rs
@@ -66,6 +66,66 @@ macro_rules! impl_neg_by_mul {
     };
 }
 
+#[cfg(test)]
+macro_rules! test_algebraic {
+    ($target:ty) => {
+        use num::Zero;
+        use approx::AbsDiffEq;
+        #[allow(unused_imports)]
+        use proptest::prelude::*;
+
+        proptest! {
+            #[test]
+            fn test_zero(a in any::<$target>()) {
+                let z = a.clone() - a;
+                prop_assert!(z.is_zero());
+            }
+
+            #[test]
+            fn test_scalar_distributive(x in any::<$target>(), y in any::<$target>(), a in -1.0..1.0_f64) {
+                let a_xy = a * (x.clone() + y.clone());
+                let ax_ay = a * x + a * y;
+                prop_assert!(a_xy.abs_diff_eq(&ax_ay, 1e-10));
+            }
+
+            #[test]
+            fn test_add_associativity(a in any::<$target>(), b in any::<$target>(), c in any::<$target>()) {
+                let ab = a.clone() + b.clone();
+                let ab_c = ab.clone() + c.clone();
+                let bc = b + c;
+                let a_bc = a + bc.clone();
+                prop_assert!(ab_c.abs_diff_eq(&a_bc, 1e-10), r#"
+                    a+b = {ab:?}
+                    b+c = {bc:?}
+                    (a+b)+c = {ab_c:?}
+                    a+(b+c) = {a_bc:?}
+                "#);
+            }
+
+            #[test]
+            fn test_mul_associativity(a in any::<$target>(), b in any::<$target>(), c in any::<$target>()) {
+                let ab = a.clone() * b.clone();
+                let ab_c = ab.clone() * c.clone();
+                let bc = b * c;
+                let a_bc = a * bc.clone();
+                prop_assert!(a_bc.abs_diff_eq(&ab_c, 1e-10), r#"
+                    a*b = {ab:?}
+                    b*c = {bc:?}
+                    (a*b)*c = {ab_c:?}
+                    a*(b*c) = {a_bc:?}
+                "#);
+            }
+
+            #[test]
+            fn test_distributive(a in any::<$target>(), b in any::<$target>(), c in any::<$target>()) {
+                let left = a.clone() * (b.clone() + c.clone());
+                let right = a.clone() * b.clone() + a * c;
+                prop_assert!(left.abs_diff_eq(&right, 1e-10));
+            }
+        }
+    };
+}
+
 mod function;
 mod linear;
 mod polynomial;
